Added user details page. Also now users can become hosts by adding information about their home.

+++ LendYourHome/LendYourHome.Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -110,7 +110,7 @@ namespace LendYourHome.Data.Migrations
                     b.Property<string>("AdditionalThoughts")
                         .IsRequired();
 
-                    b.Property<byte>("Evaluation");
+                    b.Property<int>("Evaluation");
 
                     b.Property<DateTime>("SubmitDate");
